refactor(units): move `std_ratio` concept
